Refactoring and to standart repository test
