+++ src/api/__mocks__/DELETE/v2/validation/1.json
@@ -1,0 +1,7 @@
+[
+  {
+    "response": {
+      "status": 204
+    }
+  }
+]
