+++ src/api/__mocks__/GET/v2/validations.json
@@ -1,0 +1,28 @@
+[
+  {
+    "query_string": {
+      "repository_uuid": "8511fd26-a3bc-4f74-9af1-176abca5401d"
+    },
+    "response": {
+      "status": 200,
+      "data": {
+        "count": 1,
+        "next": null,
+        "previous": null,
+        "results": [
+          {
+            "id": 1,
+            "repository_update": 1,
+            "deleted_in": null,
+            "text": "my name is douglas",
+            "intent": "greet",
+            "language": "en",
+            "created_at": "2018-05-11T11:57:40.082934Z",
+            "entities": [],
+            "translations": []
+          }
+        ]
+      }
+    }
+  }
+]
