+++ src/components/repository/repository-test/example/NewExampleTestModal.vue
@@ -1,7 +1,7 @@
 <template>
   <bh-modal
     :open.sync="openValue"
-    title="Edit sentence!">
+    title="Add a new test sentence!">
     <div class="new-example-form-modal">
       <form
         @submit.prevent="onSubmit()">
@@ -52,7 +52,7 @@
             :text-selected="textSelected"
             :available-entities="availableEntities"
             :available-labels="availableLabels"
-            :entities-for-edit="entities"
+            :entities-for-edit="[]"
             @entityAdded="onEntityAdded($event)"
             @entityEdited="onEditEntity($event)" />
         </bh-field>
@@ -64,8 +64,18 @@
             class="new-example-form-modal__buttons-wrapper__button"
             primary
             size="medium"
+            @click="submitFormAndClose()">
+            <slot v-if="!submitting">Save sentence</slot>
+          </bh-button>
+          <bh-button
+            :disabled="!isValid || submitting "
+            :tooltip-hover="!isValid ? validationErrors : null"
+            :loading="submitting"
+            primary
+            class="new-example-form-modal__buttons-wrapper__button"
+            size="medium"
             type="submit">
-            <slot v-if="!submitting">Edit your sentence</slot>
+            <slot v-if="!submitting">Save and add another</slot>
           </bh-button>
         </div>
       </form>
@@ -83,7 +93,7 @@ import { formatters } from '@/utils';
 
 
 export default {
-  name: 'EditSentenceTestModal',
+  name: 'NewExampleTestModal',
   components: {
     ExampleTextWithHighlightedEntitiesInput,
     EntitiesInput,
@@ -98,32 +108,15 @@ export default {
       type: Boolean,
       default: false,
     },
-    textToEdit: {
-      type: String,
-      default: '',
-    },
-    intentToEdit: {
-      type: String,
-      default: '',
-    },
-    entitiesToEdit: {
-      type: Array,
-      default: /* istanbul ignore next */ () => ([]),
-    },
-    sentenceId: {
-      type: Number,
-      default: null,
-    },
   },
   data() {
     return {
       textSelected: null,
-      text: this.textToEdit,
+      text: '',
       language: this.repository.language,
-      intent: this.intentToEdit,
-      entities: this.entitiesToEdit,
+      intent: '',
+      entities: [],
       errors: {},
-      id: this.sentenceId,
       submitting: false,
       openValue: this.open,
     };
@@ -178,7 +171,6 @@ export default {
     },
     data() {
       const {
-        id,
         text,
         language,
         intent,
@@ -186,7 +178,6 @@ export default {
       } = this;
 
       return {
-        id,
         text,
         language,
         intent,
@@ -204,8 +195,7 @@ export default {
   },
   methods: {
     ...mapActions([
-      'updateExampleTest',
-      'setUpdateRepository',
+      'newExampleTest',
     ]),
     setTextSelected(value) {
       this.textSelected = value;
@@ -233,7 +223,7 @@ export default {
       }
 
       try {
-        await this.updateExampleTest({
+        await this.newExampleTest({
           repository: this.repository.uuid,
           ...this.data,
         });
@@ -243,7 +233,42 @@ export default {
         this.entities = [];
         this.submitting = false;
 
-        this.setUpdateRepository(true);
+        this.$emit('created');
+        return true;
+      } catch (error) {
+        /* istanbul ignore next */
+        const data = error.response && error.response.data;
+
+        /* istanbul ignore next */
+        if (data) {
+          /* istanbul ignore next */
+          this.errors = data;
+        }
+        /* istanbul ignore next */
+        this.submitting = false;
+      }
+      return false;
+    },
+    async submitFormAndClose() {
+      this.errors = {};
+      this.submitting = true;
+      if (this.$refs.entitiesInput.clearEntityForm) {
+        this.$refs.entitiesInput.clearEntityForm();
+      }
+
+      try {
+        await this.newExampleTest({
+          repository: this.repository.uuid,
+          ...this.data,
+        });
+
+        this.text = '';
+        this.intent = '';
+        this.entities = [];
+        this.submitting = false;
+
+        this.$emit('created');
+        this.openValue = false;
         return true;
       } catch (error) {
         /* istanbul ignore next */
